+++ ios/Framework/Core/Context/SessionContext.swift
@@ -192,6 +192,9 @@ import Foundation
 		else if session.authentication is LROAuth {
 			return reloginOAuth(completed)
 		}
+		else if session.authentication is LRCookieAuthentication {
+			return reloginCookie(completed)
+		}
 
 		return false
 	}
@@ -233,8 +236,39 @@ import Foundation
 		}
 	}
 
+	open func reloginCookie(_ completed: (([String:AnyObject]?) -> Void)?) -> Bool {
+		guard session.authentication is LRCookieAuthentication else {
+			completed?(nil)
+			return false
+		}
+
+
+		SessionContext.reloadCookieAuth(session: self.session, callback: LRCookieBlockCallback { (session, error) in
+			guard session != nil, let auth = session?.authentication as? LRCookieAuthentication else {
+				print("Error reloading the cookie auth\(error!)")
+				completed?(nil)
+				return
+			}
+
+			_ = SessionContext.currentContext?.refreshUserAttributes { attributes in
+				if let attributes = attributes {
+					SessionContext.loginWithCookie(authentication: auth, userAttributes: attributes)
+				}
+				else {
+					SessionContext.logout()
+				}
+
+				completed?(attributes)
+			}
+
+		})
+
+		return true
+	}
+
 	open func refreshUserAttributes(_ completed: (([String:AnyObject]?) -> Void)?) -> Bool {
 		let session = self.createRequestSession()
+
 
 		session.callback = LRBlockCallback(
 			success: { obj in
@@ -286,15 +320,25 @@ import Foundation
 
 	@discardableResult
 	open class func loadStoredCredentials() -> Bool {
+		return loadStoredCredentials(shouldLoadServer: false)
+	}
+
+	@discardableResult
+	open class func loadStoredCredentialsAndServer() -> Bool {
+		return loadStoredCredentials(shouldLoadServer: true)
+	}
+
+	private class func loadStoredCredentials(shouldLoadServer: Bool) -> Bool {
 		guard let storage = CredentialsStorage.createFromStoredAuthType() else {
 			return false
 		}
 
-		return loadStoredCredentials(storage)
+		return loadStoredCredentials(storage, shouldLoadServer: shouldLoadServer)
 	}
 
-	open class func loadStoredCredentials(_ storage: CredentialsStorage) -> Bool {
-		guard let result = storage.load() else {
+	@discardableResult
+	open class func loadStoredCredentials(_ storage: CredentialsStorage, shouldLoadServer: Bool = false) -> Bool {
+		guard let result = storage.load(shouldLoadServer: shouldLoadServer) else {
 			return false
 		}
 		guard result.session.server != nil else {