+++ ios/Framework/Core/Context/Storage/CredentialsStore.swift
@@ -25,5 +25,6 @@ import UIKit
 	func storeCredentials(_ session: LRSession?, userAttributes: [String:AnyObject]?) -> Bool
 	func removeStoredCredentials() -> Bool
 	func loadStoredCredentials() -> Bool
+	func loadStoredCredentialsAndServer() -> Bool
 
 }